Fix sequences not properly handling concurrency
The current implementations of both SeqInt and SeqString have race
conditions that cause them to potentially produce duplicate values if
multiple objects are constructed concurrently.  In both functions the
counter variable is incremented correctly using atomic.AddInt64, but the
value is subsequently used in a line below with a raw read.  Instead,
the result of atomic.AddInt64 can simply be passed to the generator
function, ensuring proper thread-safety.

I added a test case to reproduce this behavior, and verified that it
both fails and triggers the race detector before the actual code change
to the Factory type, and succeeds after.